pyproject.toml
[project]
name = "nonebot-plugin-sendmsg-by-bots"
version = "0.0.9"
description = "a send msg tools"
authors = [
    {name = "nek0us", email = "nekouss@gmail.com"},
]
license = {text = "MIT"}
dependencies = ["nonebot2>=2.0.0", "nonebot-adapter-onebot>=2.1.3"]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/nek0us/nonebot-plugin-sendmsg-by-bots"
Repository = "https://github.com/nek0us/nonebot-plugin-sendmsg-by-bots"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"